+++ src/implementation/codegen.rs
@@ -213,13 +213,14 @@ pub(crate) fn generate(
         .iter()
         .filter(|c| c.ty == ContractType::Requires || c.ty == ContractType::Invariant)
         .flat_map(|c| {
+            let contract_type_name = if c.ty == ContractType::Invariant {
+                format!("{} (as pre-condition)", c.ty.message_name())
+            } else {
+                c.ty.message_name().to_string()
+            };
+
             let desc = if let Some(desc) = c.desc.as_ref() {
-                format!(
-                    "{} of {} violated: {}",
-                    c.ty.message_name(),
-                    func_name,
-                    desc
-                )
+                format!("{} of {} violated: {}", contract_type_name, func_name, desc)
             } else {
                 format!("{} of {} violated", c.ty.message_name(), func_name)
             };
@@ -250,13 +251,14 @@ pub(crate) fn generate(
         .iter()
         .filter(|c| c.ty == ContractType::Ensures || c.ty == ContractType::Invariant)
         .flat_map(|c| {
+            let contract_type_name = if c.ty == ContractType::Invariant {
+                format!("{} (as post-condition)", c.ty.message_name())
+            } else {
+                c.ty.message_name().to_string()
+            };
+
             let desc = if let Some(desc) = c.desc.as_ref() {
-                format!(
-                    "{} of {} violated: {}",
-                    c.ty.message_name(),
-                    func_name,
-                    desc
-                )
+                format!("{} of {} violated: {}", contract_type_name, func_name, desc)
             } else {
                 format!("{} of {} violated", c.ty.message_name(), func_name)
             };